Fix ban/mute when using weeks, months, or years
The Duration#plus() method does not support units that have an estimated
duration, which include weeks, months, and years.  This caused an
exception to be raised when any of these units were specified to a lobby
admin ban/mute action.

The fix is to convert Instant to LocalDateTime (in UTC) before adding
the ban/mute duration because LocalDateTime does support units that have
an estimated duration.