Fix broken ValidateAttachmentsTest field check
The PropertyUtil#getFieldIncludingFromSuperClasses() method only looks
for the field name as given.  It does not also try looking up a field
prefixed with "m_" as the PropertyUtil#getPropertyField() method does.

To fix this, a new getPropertyField() overload was added that accepts a
Class in place of the Object parameter in the existing overload.  The
two calls in the codebase to getFieldIncludingFromSuperClasses() were
replaced with this new overload.  The ValidateAttachmentsTest code now
runs the same logic as the production code.