Add support for virtual game properties
A virtual game property either has no backing field or uses one or more
backing fields associated with other non-virtual properties.  A virtual
property is typically used for so-called "convenience" properties that
are used to initialize multiple properties at once.  Properties of this
type are not required to have get, reset, or clear methods.