Show cursor after completing a move
The code that triggers showing the cursor after completing a move is
the call to AbstractMovePanel#cancelMove()
in AbstractMovePanel#setActive().  cancelMove() was changed in 7c6df19
to only execute its body if the cancel button is enabled.  This change
is the root cause for why the cursor is never shown after it is hidden
during a move because the cancel button has already been disabled by
the time cancelMove() is executed (due to its execution being deferred
via the call to SwingUtilities#invokeLater()).

The fix is to restore the original behavior in which the body of
cancelMove() is always executed regardless of the state of the cancel
button.  The user will still be prevented from executing this code due
to the button being disabled.  However, programmatic calls to
cancelMove() will always be executed.