Improve chat integration test
This test intermittently fails.  It appears the root cause is due to the
limited amount of time the test allows for messages to be delivered.  It
sets that limit to 250 ms, which should be more than sufficient.
However, when Travis gets heavily loaded, the time it takes to spin up
other threads to send the messages and deliver them through the network
stack can exceed this duration.  This change bumps the maximum time for
which the test may run by an order of magnitude just to be safe.  This
does not adversely affect the actual time it may take to run this test;
it simply allows more time for the expected condition to be satisfied.

Several refactorings were also applied to improve the test overall.