Use map folder override setting if present
Due to how the client settings subsystem works, the "override" will be
persisted between runs.  From the user's perspective, they probably only
expect the setting to be used for the run in which they provided the
command line argument.  To ensure this expectation is met, we have to
reset this client setting when the application starts to ensure the
normal client setting (USER_MAPS_FOLDER_PATH) will be used instead of an
override from a previous run.