Remove Derby DB bridge code
When trying to upgrade to latest we saw the bridge code did not work well. First when trying to log in, passwords were rejected. We cleared Derby database legacy data and could login with a new account, but then password challenges failed #2342

Instead of bridging accounts, to try and get the migration correct, here we just drop all of the Derby DB code and would require new accounts to be created. When rolling back lobby code, Derby data would no longer be picked up, so accounts need to be recreated anyways, and in the meantime we have no mods! Hence rolling forward successfully to restore the moderator funcionality is most important at the moment.