Fix redundant initializations and some cleanup
A redundant initialization is when we assign a value that is either the
default initialization value, or alternatively we assign a value to then
overwrite it shortly afterwards never using the original assigned value.

This is often fixed by removing the initialization value. For example:
`Object o = null` -> `Object o;`

But we often had things like this:
```
boolean result = false;
result = methodCall();
```

That would be typically cleaned up to something like:
```
boolean result = methodCall();
```

Along the way there were some cleanups that really helped, and were
partlythere to be done : )

- use for each loops instead of iterators
- reduce method visibility
- Move 'getCurrentRound()' code to GameData, avoided duplication and
fixed up some code that had the ignored initialization problem mentioned
above.
- Removed some unnecessary Box (eg: `Integer.valueOf(i)` -> `i`)
- RandomStatsDetail.java had some methods getters reduced to private, at
which point we may as well remove teh method and inline the variables.
- In ClipPlayer we returned a null Optional value! This was pretty much
a bug and would have resulted in a NPE. This was updated to return an`
Optional.empty()`