Download required assets instead of cloning entire repo
GitHub does not support conditional requests at this time.  Thus, it is
not possible to make use of the "onlyIfNewer" option of the Download
task.  Therefore, in order to prevent repeated downloads of the assets,
we set the "overwrite" option to "false".  If on the off chance an asset
is updated in the repo after it has been downloaded locally, the dev
will be required to either manually delete the individual local asset or
simply run the "clean" task in order to force the updated asset(s) to be
downloaded.