Fix thread safety bug in PlainRandomSource
PlainRandomSource uses a single MersenneTwister instance to service all
PlainRandomSource instances.  It attempts to serialize access to the
MersenneTwister, which is not thread safe, by synchronizing on the
PlainRandomSource instance lock.  If there are multiple instances of
PlainRandomSource active at once, this synchronization will not protect
against concurrent access by each of those instances to the single
MersenneTwister.

A class-level lock was introduced to protect access to the class-level
MersenneTwister instance.  Additionally, lock scope was reduced to the
minimum required.