Update lobby text, remove DB wipeout mention
When we migrated the lobby server code from Derby DB to postgres we wound up 'losing' the accounts from Derby and basically switched to a new clean database. We put up some lobby messaging to let people know they needed to recreate their account. It's been some time now, this messaging is not likely to be of help to many more people.