Add additional SwingBuilder components
* Add additional SwingBuilder components

These builders represent a subset of a larger branch. For now they are
independent of much of the existing code and with this update we add a number of
'for-now' unused methods.

The Swing builders are useful for a few things:
- functional style declaration of UI components
- type safe interfaces
- simplified APIs

In this update we add more builders for JComboBox, JCheckBox,
JTabbedPane, JSplitPane, JTextField (which has quite a number of
interesting listeners like key listener and also action listener for the
'pressed enter' event)

This update also evolves the usage of builders, instead of for example
writing:
```
add(new JLabel("label"))
```
You could now write:
```
addLabel("label")
```

The usage of GridBagLayout helper has been simplified and hidden a bit.
Client code now instead specifies a grid bag layout and a fixed column
width, components added are then behind the scenes added to the grid bag
helper. You can pretty much use a GridBag as if it were a sane layout
now : )   just add compoonents and the row count will be variable. A key
advantage to a GridBag is it fits to the components added, as opposed to
a GridLayout that stretches components to git and requires a fix number
of rows.

Of note, the composite and button attachment builders are a bit interesting in
that it combines common components that are often together and work in
tandem. These builders are also only available when building existing
components, at that point you switch builders and the 'build()' function
switches from for example a JTextField to a JComponent.

These composite components were pulled from the early setting screens and
connect to host windows.  For example notably the proxy settings window
that reveals an additional settings panel when proxy settings are on, and
hides that panel when proxy settings are off. The composite or button
attachment builders are used in-line with the builder statements and
from there switches builder on you. The idea is you would first define
the build config and then define composite builders last, eg:
```
JComponent composite = JTextFieldBuilder.builder()
  .text("text")
  .attachButton("buttonText",
      textFieldValue -> buttonAction(textFieldValue))
  .build();
```

--------------

Testing notes:
- For the most part most existing components were rewritten. The builder
code is for the most simplified versions of existing UI code, added to
support the re-written UI. While doing that work most of this code was
developed and tested manually.
- Unit test coverage should be pretty high

.

* Fix hardcoded 25 value, set to  instead.

* Create key listener function in SwingAction, use this to remove a good chunk of boiler plate code when defining swing key listener actions.

* Fix imports and opportunistic weaken a parameter type to 'Iterable'

* Remove unused imports in test code

* Fix JCombobox raw type usage, convert to 'JComboBox<String>'

* break long line