Update how_to_release.md
Update task list a bit, remove the mention to update website version, that is now automated and tracks to the latest version. Now add new process where our volunteer testers will be trying out versions and tracking results in a public spreadsheet. We need at least two independent's okay on any given version for it to to be released as a latest version.