Fix scaled map blends
When drawing a blended map tile, the renderer reuses the graphics
context from the **unscaled** relief image.  Using this graphics
context, it first draws the **scaled** relief image, and then draws the
**scaled** base image.  Thus, the resulting image is a composite of
three images:

* unscaled relief image
* scaled relief image
* scaled base image

The unscaled relief image is the root cause of the incorrect rendering
artifacts observed when the zoom is less than 100%.

The fix is to create a new empty image on which to render the scaled
relief and base images so that the unscaled relief image does not appear
in the final composite image.