Ignore command-line switches
install4j may forward command-line switches passed to the TripleA launcher
to the TripleA application itself.  For example, the "-console" switch.
At present, this switch is interpreted as a save game to initially load.
Because it is not a valid save game file name, the user is presented
with an error message.

This change simply ignores all command-line switches since TripleA
always uses the pattern "key=value" to process its command-line
arguments with the exception of a single file name or URL.