Fix NPE during TripleAFrame initialization
The TripleAFrame#showCommentLogButtonModel field must be initialized
before the TripleAMenuBar is created.  Creating the TripleAMenuBar
causes the ViewMenu to be created.  During ViewMenu creation, a call is
made to obtain TripleAFrame#showCommentLogButtonModel.  If this field is
null, the Substance L&F will throw an NPE when
JCheckBoxMenuItem#setModel() is called.

PR #2570 changed the initialization order so that
TripleAFrame#showCommentLogButtonModel was initialized after the
TripleAMenuBar was created.  This change restores the initialization
order prior to that PR such that TripleAFrame#showCommentLogButtonModel
is initialized before the creation of the TripleAMenuBar.