Allow substitution of ClientSetting Preferences implementation
Several unit tests indirectly modify one or more ClientSettings.  This
has the side effect of permanently modifying the developer's
preferences.  This can be annoying as those changes then affect the
production operation of TripleA on the developer's machine.  It also
makes the tests susceptible to changes in the developer's preferences
(e.g. enabling beta features).

This change forces all such tests to use an in-memory Preferences
implementation so that no permanent changes are made.  In order to help
developers detect indirect usage of ClientSetting within tests in the
future, the default Preferences implementation is not initialized.  If
an attempt is made to access a ClientSetting before the ClientSetting
framework has been initialized, an exception will be thrown with a
message describing this scenario.

There are two mechanisms for initializing the ClientSetting framework:

1. For production code, a call to ClientSetting#initialize() should be
made early in the application lifecycle.  This call will utilize the
default persistent Preferences implementation appropriate for the OS.

2. For test code, a call to ClientSetting#setPreferences() with an
appropriate Preferences implementation should be made before running
each test, and a call to ClientSetting#resetPreferences() should be made
after running each test.  A test fixture superclass that performs these
operations using an in-memory Preferences implementation is provided in
this change for the convenience of test authors.