Prevent excessively-long error messages when testing PBEM
Some email providers return excessively-long error messages (e.g.
thousands of characters).  Such a message causes the resulting
JOptionPane that is displayed to be several screens wide and effectively
unusable.  On some window managers, such an excessive width for a
window causes the window manager to crash (I observed this using GNOME
on Fedora 26).

We now log the full exception to the error console and provide a brief,
fixed-size message in the JOptionPane informing the user that an error
occurred.