Return serializable collection from AvailableGames#getGameNames()
AvailableGames#getGameNames() is called by
HeadlessGameServer#getAvailableGames(), which is called by getAvailableGames()
in the anonymous implementation of IServerStartupRemote stored in
ServerModel#m_serverStartupRemote.  It appears the return value of that
method is then serialized during RMI.

TreeMap$KeySet is not serializable, so we create a copy that is
serializable and return that from AvailableGames#getGameNames().