Hide GameParser Method Object details behind a facade
GameParser follows the Method Object pattern, and instances of this
class are not re-usable after parsing a game definition.  To prevent
accidental misuse of this type, the Method Object details are hidden
behind static methods that act as a facade for its usage.

The parsing methods were also renamed to better clarify the difference
between the two.  parse() completely parses a game definition, while
parseShallow() only partially parses a game definition sufficient for
use in the map list.