Add 'show new UI' button, behind beta feature flag (#2655)
* New UI first round of updates, enable new UI button as a beta feature

Functional changes: none unless "show beta features" is set to true, in which case a "show new UI button"
is shown on the main screen. Clicking that button launches a new prototype UI added here.

Things that work:
- window navigation
- single player game hosting (game host can launch game)
- multiplayer chatting

TODO:
- single player game launching (needs some simple wiring logic)
- game launching with multiple players (needs complicated wiring logic, bit difficult)
- fix up PBEM/PBF selection windows
- some more bells and whistles
- disable map selection controls for map clients

New features:
- AI selection is remembered per map

Background on new UI:
The new UI is solving two problems at once. First, we have a good bit of Swing UI code
intermixed with logic and models. Splitting the two has been an ongoing effort. A first
step was to use a static util library to build all swing components. A next evolution was
to use Builder APIs to build swing components, so far that has been successful. This update
re-writes much of the setup windows but to insteads use the swing component builder APIs.

Second, there are a number of UX problems in the setup screens that have been lingering
for some time. These problems include:
- Too many options are shown at one time, not all options are always needed. Example, map selection is ignored when joining a game.
- Flat main menu screen, overwhelming, does not allow for growth of new windows that can be opened from the main menu.

* Merge checkNotNull line

* Add some javax 'Nonnull' annotations

* Remove unused  class

* Mark interface as functional