Remove redundant null checks
The HeadlessServerSetup#m_lobbyWatcher field is final and initialized to
a non-null value.  In addition, this class is no longer Serializable,
which excludes the possibility of a null value being loaded from a
serialization stream.  Therefore, all the null checks against this field
will always be false and are thus redundant.