Fix versioned jar file name
The name of the "triplea.jar" file was changed to
"triplea-$version-all.jar" in #2729.  The code in
ClientFileSystemHelper#getRootFolder() depends on the name of the jar
file to find the root of the classpath.  The versioned jar file name
case was using the wrong format, and thus this method failed to find the
classpath root triggering an NPE.