Fix array equality check , clean 'logger.finer' statements
* Fix array equality check , clean 'logger.finer' statements

Static analysis flagged two arrays being compared against each other (equality match as opposed to logical). Turns out we do this as a way to check if both arrays are null, in both cases we have a comment saying so. This is a bit 'overly clever', we can instead dumb it down and check if both variables are not null, and remove the comment.

Along the way there were a few logger.fine statements that could get removed.

* Simplify Array equality check

* Rewrite with stream notation

* Switch findFirst to findAny