Remove System#exit() calls from PointFileReaderWriter
This class is library code and should not terminate the host application
upon an error.  The method signature was changed to allow the underlying
exception to be thrown and handled by the caller.

The application termination upon an error in the affected methods within
tool code  was preserved because it isn't clear if not exiting the
application upon an error would introduce other problems.