Remove HeadlessException handlers
The affected code is all part of one or another GUI application.  If a
HeadlessException occurs, we should allow it to terminate the thread.

In the map making tools, it's possible these were dual console/GUI
applications, and the HeadlessException handlers were hacks to
workaround running them in console mode.