Extract class to encapsulate all tool console output
Tool code shouldn't use ClientLogger because logError() doesn't operate
as expected in the separate processes used to run tools.  (In order to
use ClientLogger correctly, tools would have to initialize it similar to
what's done in GameRunner.)  In addition to ClientLogger, tool code
makes many calls to System.out and System.err to write information to
the console.

This change extracts the new ToolConsole class, which encapsulates all
access to the console for tools similar to what ClientLogger does for
the TripleA client.  All ClientLogger, System.out, and System.err usage
within tool code was replaced with ToolConsole methods.