Fix bad default map parsing loop
If we fail to parse, do not try to load default map. Loading default map calls the same method body, so we can get in a loop. The new behavior is that when there is a map parsing failure, there is an empty/game or no game selected. A very long time ago we would fall back to a hardcoded map, so this ultimately updates that original behavior to go to the empty map case instead.