Map parsing loop fix (#3036)
* Simplify factory default to not be a recursive method call

* Simplify, remove unnecessary null check, we always get a non-null value.

* Fix bad default map parsing loop

If we fail to parse, do not try to load default map. Loading default map calls the same method body, so we can get in a loop. The new behavior is that when there is a map parsing failure, there is an empty/game or no game selected. A very long time ago we would fall back to a hardcoded map, so this ultimately updates that original behavior to go to the empty map case instead.

* remove out of date comment

* Fix checkstyle and add a class javadoc

* Fix new thread, make it explicit when a new thread is spawned.

* Add not on NPE check, that is for parse failure case, not concurrency