When failing to log in to lobby (due to ban); show last few characters of hashed mac
These last 10 digits will let us know which ban row exists in DB that is blocked the ban. In case of ban mistake or bans which we want to make shorter, after the fact we have almost no way to know which user maps to which ban. By reporting the last 10 of the hashed mac, we'll be able to match up the ban in the DB.