Add readme files for future sub-projets.
The structure create here would have the lobby to client interaction be captured in a single library 'lobby-client'. Future plan is to have 'GameRunner' move to 'client', which would then in turn depend on 'lobby-client'. 'lobby-client' would start small, eventually it'll be the classes we use to talk to the lobby. With luck we'll be able to contain compatibility requirements to that library.

Last, there will be some common data model elements between client and lobby, those data models are hopefully kept to something of a minimum and would live in 'lobby-client-commonb', which both 'lobby' and 'lobby-client' would depend on.