Replace 'add.*If(boolean, Supplier<Component>)' with 'add.*If(Optional<Component>)'
With the first construct, IDE analysis really wants to highlight that the Supplier can be replaced
from a lambda, '() -> canBeNull.getSomething()' to 'canBeNull::getSomething()'

The former will be passed as a reference and would only NPE if invoked. The latter is evaluated as a parameter, which is done before the method code is executed, so we get a NPE in that case.

Instead to fix this we'll bake things into the arguments and combine the 'boolean arg + Supplier arg' to be a single 'optional' arg instead, and by using the 'optional' `map` function, we can achieve the same effect of pulling a property off of a component only when it is not null.