Fix speakerAdded NullPointerException
This is a sort-of overkill fix.
Initializing the List on decleration and adding the required keySet afterwards should already be enough to fix the NullPointerException reported in #3371 however there might be a race condition where the list is modified before all the other entries are being added (similar to the current race condition where the field is accessed before it's getting initialized), so I moved the ChannelSubscriber called further down, because this is the only call passing the chatChannelSubscribor object to somewhere.