Explicitly specify dice server port
As discussed in [1], there is a compatibility issue between the latest
and current stable clients.  #3277 set the "scheme" property value to
"https" for all MARTI servers.  This property is persisted in a save
game created by the latest pre-release, and thus read when the save game
is opened by any other client.

When the save game is opened in a stable client, the client attempts to
open an HTTPS connection to the dice server using its default port of
"80".  As this port is bound to the dice server's HTTP endpoint, the TLS
connection fails.

Alternatively, when the save game is opened in a pre-release client, the
client attempts to open an HTTPS connection to the dice server using its
default port of "-1", which instructs the HTTP stack to use the standard
default port for the specified scheme (443 in this case).  Thus, the TLS
connection succeeds.

The fix is to explicitly specify the port as 443 in the dice server
properties files.  This ensures that save games created by a pre-release
client will have the correct dice server port instead of relying on the
default port configured in code.  Thus, when such a save game is opened
by a stable client, it will behave just as if it were opened by a
pre-release client.

[1] https://forums.triplea-game.org/topic/879/civil-war-3-2-4-redrum-confederate-vs-wirkey-union-2