Fix lobby database Docker image
Not sure if this broke when the lobby-db project was introduced (#3286)
or due to backwards-incompatible changes introduced in the Postgres 9.5.x
Docker image.  Regardless, the latest Postgres 9.5 Docker image
supports direct database initialization from .sql files instead of
requiring an intermediate shell script to guide the initialization in
such a case.