Fix tooltips on casualties screen.
Previously, UnitChooserEntryIcon would override getWidth() and
getHeight() to always return a fixed size - but being placed in
a GridBag layout it would actually have a different size while
returning a fixed size. This caused problems with Swing's
TooltipManager which would still receive mouseEntered() events
based on the true size but would not show a tooltip if the
cursor position was outside of the returned getWidth().

This change makes the code only return a minimum and preferred
size, which fixes the issue. It also fixes a clipping issue for
units that are wider than 48px - such as Battleships on some
maps.