Remove 'removeListener' call
The method looks to contain a bug that was originally written in the code from the get-go, and survived replication to HeadlessServerSetup.

This problem was likely not noticed as the #remove call is not necessary. The object whose listener we are removing will go out of scope anyways removing references at that time.