Milestone: Decouple SetupPanelModel and MainPanel, remove one Observer
MainPanel has a reference to SetupPanelModel, in this update we reverse that relationship so that SetupPanelModel may refer to MainPanel but not the other way round. This means moving the callbacks so that MainPanel does not set up its own callbacks, but instead outside of MainPanel we will set it to listen to SetupPanelModel.