Copy Map maker PR template comment to triplea_maps.yaml (#3585)
The commentary/links to map makers seems most appropriate on the triplea_maps.yaml file itself. 
This comment is coming from the PR template (it probably does not belong there, if a map maker gets to a PR, they probably found some help documentation. The links instead of being removed are maybe most helpful in this file).