Decouple MainPanel from its model/controller (#3584)
* Milestone: Decouple SetupPanelModel and MainPanel, remove one Observer

MainPanel has a reference to SetupPanelModel, in this update we reverse that relationship so that SetupPanelModel may refer to MainPanel but not the other way round. This means moving the callbacks so that MainPanel does not set up its own callbacks, but instead outside of MainPanel we will set it to listen to SetupPanelModel.

* Cleanup update, add a comment and move flow conditional logic to caller

* Move MainPanel to its own package

* Use a named interface instead of #Consumer: 1. better naming, 2. works better with tooling when looking up usage

* Move GameSelectorModel observer attachment to MainPanel out from MainPanel constructor

* Introduce a MainPanelBuilder to remove UI construction code from GameRunner

* Move #GameSelectorPanel, turns out it is "owned" by #MainPanel. (Also included some cleanup/javadoc changes)

* Fix assignable from check by adding a method to get the drawable component.

The assignable check seems to be a way to protect the cast that happens on the next line. Instead by using a method we can avoid the cast and we will throw an exception for the headless case where we do not expect a component.

* Correct import orderings, cleanup and add a javadoc (fix checkstyles)

* Move observer binding to outside of constructor

* Move #getFormatted..Text method to model. In part the model is a view-model, in other part cleaning up GameSelectorPanel at all is useful as it contains a bit too much.

* GameSelectorPanel: Inline methods and combine #updateGameData with #setWidgetActivation

* Move GameSelector to its own package in anticipation of further refactor to break-up the class to smaller pieces.