Fix race condition in RemoteMessengerTest that can cause deadlock (#3603)
A single semaphore was used to signal two independent events in the
testMethodReturnsOnWait() test case.  This introduced a race condition
in which the test thread may not ever see the semaphore with a non-zero
permit count because the background thread immediately attempts to
re-acquire the semaphore after releasing it.  In this case, the test
thread will block forever attempting to acquire the semaphore held by
the background thread.

Fixed by introducing two separate latches to represent the two
independent signals.