Decouple lobby integration tests from file system
The lobby integration tests rely on a copy of lobby.properties under the
game-core subproject, whereas the official lobby.properties is under the
lobby subproject.  Having two files creates a bit of confusion for
developers.

In addition, the lobby code has a hidden dependency on lobby.properties
via the LobbyContext global utility.

This change removes the LobbyContext global dependency and replaces it
with an explicit dependency on LobbyPropertyReader injected into all
dependents, where required.  This then decouples the lobby integration
tests from the file system; an in-memory LobbyPropertyReader is used for
all tests instead.

LobbyRunner (the entry point of the lobby server) is now the only entity
that reads lobby.properties from the file system.